More work on image uploads

+++ code/app/controllers/FilesCtrl.scala
@@ -1,5 +1,6 @@
 package controllers
 
+import java.nio.file.attribute.PosixFilePermission._
 import java.nio.file.{Files, Paths}
 import java.sql.Timestamp
 import java.util.Calendar
@@ -12,6 +13,7 @@ import models.KMImage
 import play.api.libs.json.{JsError, Json}
 import play.api.mvc.{ControllerComponents, InjectedController, PlayBodyParsers}
 import play.api.{Configuration, Logger}
+
 import scala.collection.JavaConversions._
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -23,27 +25,41 @@ class FilesCtrl @Inject() (images:ImagesDAO, cc:ControllerComponents, parsers:Pl
   def apiAddFile(subjectId: String) = deadbolt.SubjectPresent()(cc.parsers.multipartFormData ){ req =>
     val uploadedFile = req.body.files.head
     val filePath = Paths.get(config.get[String]("hear_us.files.km.folder")).resolve(subjectId)
-    Files.createDirectories( filePath )
+    if ( ! Files.exists(filePath) ) {
+      Files.createDirectories( filePath )
+      Files.setPosixFilePermissions(filePath, Set(OWNER_READ, OWNER_WRITE, OWNER_EXECUTE, GROUP_READ, GROUP_EXECUTE, OTHERS_READ) )
+    }
     val filename = req.body.dataParts("qqfilename").head
     val suffix = getFileSuffix(filename)
-    images.addFile( KMImage(0L, subjectId.toLong,  suffix, "", new Timestamp(Calendar.getInstance().getTime().getTime()), getFileName(filename)) )
-      .map( f => {
-        val localPath = filePath.resolve(f.id + "." + suffix)
-        uploadedFile.ref.moveTo(localPath , replace = true)
-        import java.nio.file.attribute.PosixFilePermission._
+    
+    Logger.info( s"got file $filename with suffix $suffix")
+    Logger.info( "Uploaded: " + uploadedFile.toString  + " -- " + uploadedFile.ref.toString )
+    
+    images.addFile( KMImage(0L, subjectId.toLong,  suffix, uploadedFile.contentType.getOrElse(""), new Timestamp(Calendar.getInstance().getTime.getTime), getFileName(filename)) )
+      .map(fileDbRecord => {
+        val localPath = filePath.resolve(fileDbRecord.id + "." + suffix)
+        
+        Logger.info("Moving file to " + localPath.toAbsolutePath.toString )
+        val res = uploadedFile.ref.atomicMoveWithFallback(localPath)
+        Logger.info( "res: " + res.toString)
+        
         Files.setPosixFilePermissions(localPath, Set(OWNER_READ, OWNER_WRITE, GROUP_READ, OTHERS_READ) )
-        Ok(Json.toJson(f))
+        // top object must contain a "success:true" field, for fineUploader to know the upload went well.
+        Ok(Json.obj( "success"->true, "record"->Json.toJson(fileDbRecord)) )
       })
 
   }
 
   def deleteFile(fileId: Long) = deadbolt.SubjectPresent()() { req =>
+    Logger.info("Deleting file "  + fileId )
     images.getFile(fileId).flatMap {
       case None => Future(NotFound("file not found"))
       case Some(file) => {
         images.deleteFile(file).map(count => {
           val filePath = Paths.get(config.get[String]("hear_us.files.km.folder")).resolve(file.kmId.toString).resolve(fileId.toString + "." + file.suffix)
-          Files.delete(filePath)
+          if ( Files.exists(filePath) ) {
+            Files.delete(filePath)
+          }
           Ok(Json.toJson(count))
         })
       }
@@ -68,16 +84,12 @@ class FilesCtrl @Inject() (images:ImagesDAO, cc:ControllerComponents, parsers:Pl
   }
 
   private def getFileSuffix (fileName:String): String = {
-    var extension = ""
     val i = fileName.lastIndexOf('.')
-    if (i > 0) extension = fileName.substring(i + 1)
-    extension
+    if (i >= 0) fileName.substring(i + 1) else ""
   }
 
   private def getFileName (fileName:String): String = {
-    var name = ""
     val i = fileName.lastIndexOf('.')
-    if (i > 0) name = fileName.substring(0,i)
-    name
+    if (i > 0) fileName.substring(0,i) else fileName
   }
 }
